Create hexagonal grid

Create a hexagonal grid from an sf object.

Usage

iso_grid(x, resolution = 9)

Arguments

x

An sf object the convex hull of which will determine the hex grid

resolution

The resolution of the hex grid

Value

An sfc object

Details

See the documentation of the upstream h3 library for more information. Note: default value of 9 represents around 0.1 km^2, you may way to consider higher values (e.g. 10, which is around 0.015 km^2 or 15k m^2) for more detailed analysis or lower values for larger areas.

Note

This function is a wrapper for h3o::sfc_to_cells and h3o::flatten_h3. It assumes that the input object has a geographic (lon/lat) coordinate system, e.g. EPSG:4326.

Examples

x = net_oldenburg_raw
iso_grid(x)
#> Geometry set for 48 features 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 8.195993 ymin: 53.12823 xmax: 8.23453 ymax: 53.14946
#> Geodetic CRS:  WGS 84
#> First 5 geometries:
#> POLYGON ((8.195993 53.14152, 8.196243 53.13993,...
#> POLYGON ((8.198845 53.13939, 8.199095 53.1378, ...
#> POLYGON ((8.200947 53.14204, 8.201197 53.14045,...
#> POLYGON ((8.198095 53.14417, 8.198345 53.14258,...
#> POLYGON ((8.196743 53.13674, 8.196993 53.13515,...